fix: correct 64-bit integer types on x86

+++ include/stdint.h
@@ -1,24 +1,24 @@
 #ifndef _STDINT_H
 #define _STDINT_H
 
-typedef signed char	int8_t;
+typedef signed char	    int8_t;
 typedef unsigned char	uint8_t;
 
 typedef signed short	int16_t;
 typedef unsigned short	uint16_t;
 
-typedef signed int	int32_t;
+typedef signed int	    int32_t;
 typedef unsigned int	uint32_t;
 
-typedef signed long	int64_t;
-typedef unsigned long	uint64_t;
+typedef signed long	long      int64_t;
+typedef unsigned long long    uint64_t;
 
 #if defined(__x86_64__) || defined(__x86_64)
 typedef uint64_t	uintptr_t;
-typedef int64_t	intptr_t;
+typedef int64_t	    intptr_t;
 #else
 typedef uint32_t	uintptr_t;
-typedef int32_t	intptr_t;
+typedef int32_t	    intptr_t;
 #endif
 
 #endif
